chore(GIFT-24669): markups
