--- src/modules/gift/helpers/has-obligations-with-maturity-date-following-facility/index.ts
@@ -1,11 +1,0 @@
-type GiftFacilityObligation = {
-  maturityDateFollowsFacility?: boolean;
-};
-
-/**
- * Checks whether any obligation in a facility has maturityDateFollowsFacility set to true.
- * @param {GiftFacilityObligation[]} obligations: Facility obligations.
- * @returns {boolean} True when at least one obligation follows the facility maturity date.
- */
-export const hasObligationsWithMaturityDateFollowingFacility = (obligations: GiftFacilityObligation[]): boolean =>
-  Array.isArray(obligations) && obligations.some(({ maturityDateFollowsFacility }) => maturityDateFollowsFacility === true);
